requirements.txt
tls_client
beautifulsoup4
pycryptodome
tqdm
gdown